Fixing map component
